requirements.txt
streamlit
google-generativeai
tqdm
IPython
json2html